Add source files in the location where progen will see them and add them to the project.
Note: for the stm32h743 it appears to need to be created as a
uvision5 project, when trying with uvision (i.e. uvision4) it had an error as this part
is not in the legacy pack.

progen  generate -f projects.yaml -p stm32h743ii_if -t uvision5

NOTE: There needs to be a stm32h743 mcu definition file added to the "project_generator_definitions" tool before progen will succeed.
This is not yet in the project_generator_definitions github project, but this stm32h743iikx.yaml file can be created in the local copy of the
python package, which is typically located in this path (relative to the DAPLink project root):
  .\venv\Lib\site-packages\project_generator_definitions\mcu\st\

To set this up, create a file named stm32h743iikx.yaml in the location listed above, and
paste the following text (the part between the #### separators) into it.
Then progen should run.
####
mcu:
    vendor:
        - st
    name:
        - stm32h743iikx
    core:
        - cortex-m7
tool_specific:
    iar:
        OGChipSelectEditMenu:
          state:
          - STM32H743iikx ST STM32H743iikx
        OGCoreOrChip:
          state:
          - 1
    uvision:
        TargetOption:
            Cpu:
            #- IROM(0x08000000-0x080FFFFF) IRAM(0x20000000-0x2004FFFF) CPUTYPE("Cortex-M7") FPU3(SFPU) CLOCK(12000000) ELITTLE
            - IRAM(0x20000000,0x00020000) IRAM2(0x24000000,0x00080000) IROM(0x08000000,0x00100000) IROM2(0x08100000,0x00100000) XRAM(0x30000000,0x00048000) XRAM2(0x38000000,0x00010000) CPUTYPE("Cortex-M7") FPU3(DFPU) CLOCK(12000000) ELITTLE
            Device:
            - STM32H743IIKx
            DeviceId:
            - 0
            FlashDriverDll:
            - UL2CM3(-S0 -C0 -P0 -FD20000000 -FC8000 -FN1 -FF0STM32H7x_2048 -FS08000000 -FL0200000 -FP0($$Device:STM32H743IIKx$CMSIS\Flash\STM32H7x_2048.FLM))
            SFDFile:
            - $$Device:STM32H743IIKx$CMSIS\SVD\STM32H743.svd
            Vendor:
            - STMicroelectronics
    uvision5:
        TargetOption:
            Cpu:
            - IRAM(0x20000000,0x00020000) IRAM2(0x24000000,0x00080000) IROM(0x08000000,0x00100000) IROM2(0x08100000,0x00100000) XRAM(0x30000000,0x00048000) XRAM2(0x38000000,0x00010000) CPUTYPE("Cortex-M7") FPU3(DFPU) CLOCK(12000000) ELITTLE
            Device:
            - STM32H743IIKx
            DeviceId:
            - 0
            FlashDriverDll:
            - UL2CM3(-S0 -C0 -P0 -FD20000000 -FC8000 -FN1 -FF0STM32H7x_2048 -FS08000000 -FL0200000 -FP0($$Device:STM32H743IIKx$CMSIS\Flash\STM32H7x_2048.FLM))
            PackID:
            - Keil.STM32H7xx_DFP.2.5.0
            SFDFile:
            - $$Device:STM32H743IIKx$CMSIS\SVD\STM32H743.svd
            RegisterFile:
            - $$Device:STM32H743IIKx$Drivers\CMSIS\Device\ST\STM32H7xx\Include\stm32h7xx.h
            Vendor:
            - STMicroelectronics
####

A second quirk is that the C/C++ and Asm flags should be edited, to change
  DAPLINK_IF
to
  DAPLINK_BL
otherwise the DAPLink USB drive will not appear.  The root cause of that is still being looked at.