Fix LED blink duration for UART activity b/193370721
The LED was not visibly blinking for UART activity.

It looks like
  68eeee4dadb46d312fc13b23e0851dcec9cc9b49
  Mathias Brossard, 5 months ago   (February 9th, 2021 4:48pm)
  Fix for RTOS2 configuration (problem spotted by @gerargz)
fixed this in upstream by adding an OS_TICK_FREQ compilation parameter (set to 100Hz = 10msec tick).
As we are using a makefile vs. yaml files + project generator we need to add to our makefile.

With this change a visible LED blink is seen, and scope measurements show 30 msec pulsewidth.