Fix a hang seen in the UART IRS when the target was reset while the CDC_UART was being closed. b/193392556
The USART->ISR ORE (overrun) interrupt was being set, and nothing was
clearing it, so the interrupt kept firing.

This change will prevent some other (unhandled) interrupts from being raised,
and it clears the overrun interrupt.
In the future we might add counters to monitor for overruns / other UART errors.