[UDB LOG] provide logging functionality for debugging
Create a RAM FIFO to keep log info. If the cdc b is not connected, the log
goes to FIFO. If the cdc b is connected, we flush the log to the
console. This implementation relies on terminal's flow control. But we
also can detect cdc connection when user type the first character.

Issue: b/200237296
Change-Id: I749d59bdf3fa012d2ec7d0224c575c5fd4e0aba5