Reset usb bulk driver if openocd is trying to disconnect
Packet losses can put this driver into a bad state. I.e. if a transfer
complete ACK gets lost and USBD_BULK_EP_BULKIN_Event isn't called,
USB_ResponseIdle won't be set to get ready for the next transfer. In
that state DAPLink will respond to the current  DAP command with the
previous DAP command's response.  In such cases, openOCD will try to
disconnect and connect again. As a workaround, when   the disconnect
command is received, reset this driver and clear the TX buffer.

Issue: b/208896670
Change-Id: I8f280e1fa74fd75457c57079697649990f4b36ec