[UDB MSC] Fix bootloader mass storage
Change TX2_FIFO_SIZE back to 512.

TX2_FIFO_SIZE is the FIFO for MSC. If TX2_FIFO_SIZE is less
than the max packet size of MSC, it would get stuck in USBD_WriteEP().
For some reason I don't know, if I set msc max packet size below 512,
the host would send msc reset command and can not recover from the
reset. The only way I found is to set TX2_FIFO_SIZE back to 512,
and change the cdc b packet size to 448 so that the overall FIFO for
USB is still within the limit of 4096.

Issue: b/221158171
Change-Id: Id8fef43189782f307ebb39417cfa7338a019605d