Update Jenkins to build for "freertos" and with GCC 7.3.1
Seems that our build infrastructure will use the "platform" argument to
decide which type of host to build on. When set to "linux", it builds on
some Debian host with some sort of janky arm-none-eabi-gcc from a folder
"/tmp/building". Instead of doing that, switch to "freertos", which will
build in the usual embedded environment and includes all the toolchains
we use for other projects. I chose GCC 7.3.1, since I think that's the
latest one available (I wonder if this can be updated).

bug: 222442977
Change-Id: I4fa3406c223e3f40625a6312a26480fca139a76b