Increase stack to 2048 (from 1024).
For UDB (on stm32H743) we use AXI SRAM, which has 512KB in total.
And UDB only uses approximately 145KB, including 1KB of main_task stack.
We hit the stack overflow check when adding several recent features, so
increase it by a large margin to avoid this going forward.

b/218777300

Change-Id: Idfdc57cddc7f5daba44f698162f3e23df251fe9c