Separate UDB from stm32h743 in project.yaml
Just a small cleanup task that should make our lives easier when we add
macros and when we get to opensourcing the project.

In the past we've been adding UDB functionality to the generic stm32h743
yaml. Ideally the two should be able to live separately, because from
DAPLink's perspective not all stm32h743ii is a UDB.

Change-Id: I04bf1b980fdd3e719548905e1c8ea58e50034cc6