Remove BL version from welcome message
The bootloader is updated after the prerun_board_config, meaning the
welcome message has the old BL version after an update. If the device is
reset, the next welcome message has the correct version. Might be
confusing for devs and users, and the only decent solution I could find
is to force users to manually read the BL version.

Change-Id: Ib0e28164929d55459cc90abf59ee9a9e337aeab9