Print bootloader version in welcome msg
Previously removed it from because the welcome msg was printed in
the prerun_board_config, which ran before the bootloader update.
At that time I didn't want to add an if statement into the 30_ms hook
but no we have another user for that if statement, so we can use it for
the welcome message too.

Change-Id: I342055392fbb2759347459ea0592e920df1c9a30