Fix up integration test
The voltage and current limits were too tight and the test was failing
on my P2 UDB.

The watchdog test requires the user to unplug the debugger, otherwise
UDB will halt when the watchdog reset happens. Skip the test unless the
user specifically wants to run it.

Depending on timing, opening the serial port right after a reset will
give different errors. Added OSError in a few places to cover this
variability.

Change-Id: Ic8dd63b09895f9cf26264d76ed50ca3032a5612d