[UDB USB] Fix USB hang issue when using many endpoints simultaneously
In ReadEP(), after we process the data in Rx FIFO, we would
wait for the transfer complete (XFRC) or setup done (STUP) packet and
pop it out. However, when lots of traffic go in udb, sometimes a setup
packet would go into the FIFO before the core pushes XFRC/STUP, which
would lead to a disaster. Verify it is indeed XFRC/STUP and then pop it
out.

If the host stops sending IN token after we push data into the Tx FIFO,
the next IN packet would wait infinitly until the watchdog fires.
Add a timeout to limit the waiting time, and if we can not get enough space
in time, reset the IN endpoint.

Issue: b/205946697
Change-Id: I75fcdbd14e1d0225e7108fdfbe1608898238174f